Take into account neutral units for walkable tiles and pathfinding. Invalidate pathfinding cache when a neutral unit dies. (Also, there are probably bugs now because paths can be null now and we definitely don't check that)

+++ Bot/MapKnowledge/ExpandAnalyzer.cs
@@ -17,7 +17,7 @@ public class ExpandAnalyzer: INeedUpdating {
     public static List<List<Unit>> ResourceClusters;
     public static List<Vector3> ExpandLocations;
 
-    private const float ExpandSearchRadius = 5;
+    private const int ExpandSearchRadius = 5;
 
     private ExpandAnalyzer() {}
